+++ infra/resources/modules/session_manager/locals.tf
@@ -29,5 +29,21 @@ locals {
     IO_PROFILE_API_BASE_PATH = var.io_profile.base_path
     # IO_PROFILE_API_KEY       = "@Microsoft.KeyVault(SecretUri=${azurerm_key_vault_secret.sm_io_profile_api_key.versionless_id})"
     IO_PROFILE_API_KEY = "TODO"
+
+    # Fast Login service
+    IO_FAST_LOGIN_API_URL       = var.io_fast_login.base_url
+    IO_FAST_LOGIN_API_BASE_PATH = var.io_fast_login.base_path
+    # IO_FAST_LOGIN_API_KEY       = "@Microsoft.KeyVault(SecretUri=${azurerm_key_vault_secret.sm_io_fast_login_api_key.versionless_id})"
+    IO_FAST_LOGIN_API_KEY = "TODO"
+
+    # Locked Profiles table
+    LOCKED_PROFILES_STORAGE_ACCOUNT_URI = "https://${var.locked_profiles.storage_account.name}.table.core.windows.net"
+    LOCKED_PROFILES_TABLE_NAME          = var.locked_profiles.table_name
+
+    # Session Manager Internal service
+    IO_SM_INT_API_URL   = var.io_session_manager_internal.base_url
+    IO_SM_INT_BASE_PATH = var.io_session_manager_internal.base_path
+    # IO_SM_INT_API_KEY       = "@Microsoft.KeyVault(SecretUri=${azurerm_key_vault_secret.sm_io_sm_int_api_key.versionless_id})"
+    IO_SM_INT_API_KEY = "TODO"
   }
 }