+++ packages/io-auth-n-identity-domain/src/domain/value-objects/lollipop/lollipop-assertion-type.vo.ts
@@ -1,0 +1,4 @@
+import { z } from "zod";
+
+export const LollipopAssertionTypeSchema = z.enum(["SAML", "OIDC"]);
+export type LollipopAssertionType = z.infer<typeof LollipopAssertionTypeSchema>;